Extend the layer for docker's default behavior to set an intermediate state
This has the intended effect of allowing workloads that need to modify the
docker daemon to do so before any workloads are launched. This should help
alleviate cycling production services and --restart policy woes, but its
not fool proof. For example if you are waiting on an extant service to
become joined, the docker.available state will raise regardless.