Includes service wrappers for Docker, DockerCompose
This is a breakout of lib concerns for the Docker charm base, which includes
a CLI compatible wrapper for executing docker commands in given contexts.
This is all convenience for the charm author - instead of littering
subprocess.check_call or subprocess.check_output methods in their hook
code. This abstraction allows for a clean API to perform common tasks.

Example Usage - use the bootstrap docker daemon to start flannel:

	from charms.docker import Docker
	d = Docker(socket="unix:///var/run/docker-bootstrap.sock")
	if d.running():
	    d.run('quay.io/flannel/flannel', ports="4001:4001")
	    hookenv.open_port(4001)

Example Usage for DockerCompose - spin up nginx out of a composer formation

	from charms.docker.compose import Compose
	c = Compose('files/nginx-webhead')
	c.start_service('nginx')

The modules are further documented in the source files.

All the modules are tested via `tox` - after cloning the repository

    pip install tox
    tox