Adds the skip_install layer option
This is useful when building layer-docker based subordinates. The charm
has a self-relationship that is capable of denoting we are connecting
to an already juju maintained version of docker on the host. Why sit
through the implicit 40 sconds of probing when we can skip and get
directly to the workload.

Default behavior is to run the install hook though.