Encapsulate docker-compose in the wheelhouse
Docker-compose has been pulled from the upstream docker ppa, which was weird, as it existed for a couple weeks. Regardless, installing from PIP on the production system is iffy. I think installing from the wheelhouse and keeping docker-compose encapsulated in the charm is the best path forward for now.

This commit also rolls up a few cleanup operations around the charm such as updating the readme, and adding a repo key to layer.yaml