Stuff the docker components in the wheelhouse
:celebrate: Finally, some real results with old work. Supercedes  #41
and removes the hacky fixes in #65.

What this is doing:

 - We needed a tactic to allow us to modify the virtualenv thats packing
   the wheelhouse. This allowed us to update components for that venv,
   which resulted in a winning wheelhouse package!

Now we can offline dockerpy and docker-compose. This tactic may be
useful to other people, which we may deprecate this and rename it and
stuff it in the builder proper. I'd like more feedback on this before
proposing and let us kick the tires for a bit while we figure out if its
doing anything really unholy in practice.