Defaults to archive
Until this commit, layer-docker defaulted to installing from the docker
ppa. Docker 1.12 introduces a lot of interesting behavior changes behind
the scenes that can be problematic. Until we have a better pinning story
with this layer, I feel that this is the best path forward.

This will give our users the choice of which "distribution" of docker
they wish to install, and gives us the capacity to ensure we're building
against the work being put forth in the ubuntu archive. (read:
deployment into lxd containers with the docker profile)